backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pinecone==3.2.2
google-generativeai==0.3.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-multipart==0.0.6
sse-starlette==1.8.2
